refactor(aws): fold logs_client.go into client.go
Keep the CloudWatch alarm/metric Client and the CloudWatch Logs
LogsClient as distinct types (different wire protocols and
endpoints) but in one file, matching the rest of the package's
single client.go convention.
